+++ eva/src/components/autorization/mainPage.vue
@@ -27,23 +27,6 @@
           <v-card
             class="product-block-itself"
             :style="{ backgroundColor: theme.$main_bg }"
-            @click="openReport"
-          >
-            <v-card-text
-              class="product-text"
-              :style="{ color: theme.$title }"
-            >
-              Исследование <br>
-              данных
-              <p
-                class="delimetr"
-                :style="{ backgroundColor: theme.$title }"
-              />
-            </v-card-text>
-          </v-card>
-          <v-card
-            class="product-block-itself"
-            :style="{ backgroundColor: theme.$main_bg }"
             @click="openDataResearch"
           >
             <v-card-text
@@ -75,23 +58,6 @@
               />
             </v-card-text>
           </v-card>
-          <v-card
-            class="product-block-itself"
-            :style="{ backgroundColor: theme.$main_bg }"
-            @click="openPaper"
-          >
-            <v-card-text
-              class="product-text"
-              :style="{ color: theme.$title }"
-            >
-              Конструирование <br>
-              отчетов
-              <p
-                class="delimetr"
-                :style="{ backgroundColor: theme.$title }"
-              />
-            </v-card-text>
-          </v-card>
         </div>
       </v-container>
     </v-main>
@@ -116,16 +82,8 @@ export default {
     openDash() {
       this.$router.push('/dashboards');
     },
-    openReport() {
-      this.$store.commit('createReportSearch');
-      this.$router.push('/reports');
-    },
     openDataResearch() {
       this.$router.push('/data-research');
-    },
-    openPaper() {
-      this.$store.commit('createPaperSearch');
-      this.$router.push('/papers');
     },
     openQuis() {
       window.location.href = '/quiz/list'; // используем window.location.href потому что /quiz/list внешняя ссылка а не часть приложения