+++ CHANGELOG.md
@@ -12,6 +12,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 - Fix accumulator colors
+- Fixed title actions
+- Removed old reports page 
+- Removed papers page
+- Remove title actions from reports visualisations
 
 ## [1.91.0]